# Increase performance and decrease memory usage