@janlelis 2016-12-09
Hey rrosenblum,

thank you for investigating and opening the PR, will totally merge this. Btw, which tool did you use for memory profiling?