@rrosenblum 2016-12-09
@janlelis, the gem that I used for memory profiling is [`memory_profiler`](https://github.com/SamSaffron/memory_profiler).